+++ src/core/prelude/number/metrics/index.ts
@@ -8,6 +8,14 @@
 
 import extend from 'core/prelude/extend';
 
+/** @see [[Number.isSafe]] */
+extend(Number.prototype, 'isSafe', function isSafe(this: number): boolean {
+	return this >= Number.MIN_SAFE_INTEGER && this <= Number.MAX_SAFE_INTEGER;
+});
+
+/** @see [[NumberConstructor.isSafe]] */
+extend(Number, 'isSafe', (value) => Object.isNumber(value) && value.isSafe());
+
 /** @see [[Number.isInteger]] */
 extend(Number.prototype, 'isInteger', function isInteger(this: number): boolean {
 	return this % 1 === 0;