+++ src/core/prelude/structures/sync-promise/spec.js
@@ -22,6 +22,92 @@ describe('core/prelude/structures/sync-promise', () => {
 		expect(i).toBe(6);
 	});
 
+	it('promise that is resolved with another promise', async () => {
+		const i = await new SyncPromise((resolve) => {
+			resolve(
+				new Promise((r) => setTimeout(() => r(SyncPromise.resolve(1)), 50))
+			);
+		})
+			.then((val) => new Promise((r) => setTimeout(() => r(SyncPromise.resolve(val + 2)), 50)))
+			.then((val) => val * 2);
+
+		expect(i).toBe(6);
+	});
+
+	it('promise that is rejected with another promise', async () => {
+		let i;
+
+		try {
+			await new SyncPromise((resolve) => {
+				resolve(
+					new Promise((r) => setTimeout(() => r(SyncPromise.resolve(1)), 50))
+				);
+			})
+				.then((val) => new Promise((r) => setTimeout(() => r(SyncPromise.reject(val + 2)), 50)))
+				.catch((val) => Promise.reject(val * 2));
+
+		} catch (err) {
+			i = err;
+		}
+
+		expect(i).toBe(6);
+	});
+
+	it('promise that is rejected with another promise by using a constructor', async () => {
+		let i;
+
+		try {
+			await new SyncPromise((resolve, reject) => {
+				reject(SyncPromise.resolve(1));
+			});
+
+		} catch (err) {
+			i = err;
+		}
+
+		expect(i).toBeInstanceOf(SyncPromise);
+		expect(await i).toBe(1);
+	});
+
+	it('double promise resolution', async () => {
+		expect(
+			await new SyncPromise((resolve) => {
+				resolve(1);
+				resolve(2);
+			})
+		).toBe(1);
+
+		expect(
+			await new SyncPromise((resolve) => {
+				resolve(new Promise((r) => setTimeout(() => r(1)), 100));
+				resolve(2);
+			})
+		).toBe(1);
+	});
+
+	it('double promise rejection', async () => {
+		try {
+			await new SyncPromise((resolve, reject) => {
+				reject(1);
+				reject(2);
+			});
+
+		} catch (err) {
+			expect(err).toBe(1);
+		}
+
+		try {
+			await new SyncPromise((resolve, reject) => {
+				reject(new Promise((r) => setTimeout(() => r(1)), 100));
+				reject(2);
+			});
+
+		} catch (err) {
+			expect(err).toBeInstanceOf(Promise);
+			expect(await err).toBe(1);
+		}
+	});
+
 	it('resolved `then` after `catch`', () => {
 		let
 			i = 1;
@@ -179,6 +265,40 @@ describe('core/prelude/structures/sync-promise', () => {
 		expect(i).toBe(6);
 	});
 
+	it('`finally` that returns an error', () => {
+		let
+			reason;
+
+		new SyncPromise((resolve) => {
+			resolve(1);
+		})
+			.finally(() => SyncPromise.reject('Boom'))
+
+			.catch((err) => {
+				reason = err;
+			});
+
+		expect(reason).toBe('Boom');
+	});
+
+	it('`finally` that throws an error', () => {
+		let
+			reason;
+
+		new SyncPromise((resolve) => {
+			resolve(1);
+		})
+			.finally(() => {
+				throw 'Boom';
+			})
+
+			.catch((err) => {
+				reason = err;
+			});
+
+		expect(reason).toBe('Boom');
+	});
+
 	it('`SyncPromise.resolve`', () => {
 		let
 			i = 0,
@@ -214,11 +334,8 @@ describe('core/prelude/structures/sync-promise', () => {
 			let
 				res;
 
-			SyncPromise.all([
-				1,
-				null,
-				SyncPromise.resolve(2)
-			]).then((val) => res = val);
+			SyncPromise.all([1, null, SyncPromise.resolve(2)])
+				.then((val) => res = val);
 
 			expect(res).toEqual([1, null, 2]);
 		});
@@ -227,11 +344,7 @@ describe('core/prelude/structures/sync-promise', () => {
 			let
 				res;
 
-			SyncPromise.all([
-				1,
-				null,
-				SyncPromise.reject(2)
-			]).then(
+			SyncPromise.all([1, null, SyncPromise.reject(2)]).then(
 				(val) => res = val,
 				(err) => res = err
 			);
@@ -244,11 +357,7 @@ describe('core/prelude/structures/sync-promise', () => {
 		let
 			res;
 
-		SyncPromise.allSettled([
-			1,
-			null,
-			SyncPromise.reject(2)
-		]).then(
+		SyncPromise.allSettled([1, null, SyncPromise.reject(2)]).then(
 			(val) => res = val,
 			(err) => res = err
 		);
@@ -265,10 +374,8 @@ describe('core/prelude/structures/sync-promise', () => {
 			let
 				res;
 
-			SyncPromise.race([
-				Promise.resolve(1),
-				SyncPromise.resolve(2)
-			]).then((val) => res = val);
+			SyncPromise.race([Promise.resolve(1), SyncPromise.resolve(2)])
+				.then((val) => res = val);
 
 			expect(res).toBe(2);
 		});
@@ -277,10 +384,7 @@ describe('core/prelude/structures/sync-promise', () => {
 			let
 				res;
 
-			SyncPromise.race([
-				Promise.resolve(1),
-				SyncPromise.reject(2)
-			]).then(
+			SyncPromise.race([Promise.resolve(1), SyncPromise.reject(2)]).then(
 				(val) => res = val,
 				(err) => res = ['error', err]
 			);