+++ src/core/prelude/number/metrics/CHANGELOG.md
@@ -9,6 +9,12 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.60.0 (2021-10-01)
+
+#### :rocket: New Feature
+
+* Added `Number.isSafe` and `Number.prototype.isSafe`
+
 ## v3.20.0 (2020-07-05)
 
 #### :rocket: New Feature