+++ src/core/promise/sync/CHANGELOG.md
@@ -9,11 +9,27 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
-## v3.?.? (2021-??-??)
+## v3.60.3 (2021-10-05)
+
+#### :bug: Bug Fix
+
+* Fixed a bug with double resolving of a promise when it resolved by another promise
+
+## v3.59.2 (2021-09-27)
+
+#### :bug: Bug Fix
+
+* Now `finally` respects exceptions values from a callback
+
+## v3.59.0 (2021-09-08)
 
 #### :rocket: New Feature
 
 * Added a new static method `any`
+
+#### :bug: Bug Fix
+
+* Fixed some cases when a promise is resolved with another promise
 
 ## v3.58.0 (2021-09-01)